+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtoolcore.compare.diff_identifiers`` helper function
+- ``dtoolcore.compare.diff_sizes`` helper function
+- ``dtoolcore.compare.diff_content`` helper function
+
 
 Changed
 ^^^^^^^